requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
Pillow
tqdm
scikit-learn
pycocotools